+++ Source/OxyPlot.WindowsForms/PngExporter.cs
@@ -90,9 +90,11 @@ namespace OxyPlot.WindowsForms
             var bm = new Bitmap(this.Width, this.Height);
             using (var g = Graphics.FromImage(bm))
             {
-                if (this.Background.IsVisible())
+                var actualBackground = this.Background.IsUndefined() ? model.Background : this.Background;
+
+                if (actualBackground.IsVisible())
                 {
-                    using (var brush = this.Background.ToBrush())
+                    using (var brush = actualBackground.ToBrush())
                     {
                         g.FillRectangle(brush, 0, 0, this.Width, this.Height);
                     }